+++ sentry-kotlin-multiplatform/src/appleMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryEventExtensions.apple.kt
@@ -5,12 +5,17 @@ import io.sentry.kotlin.multiplatform.CocoaSentryEvent
 import io.sentry.kotlin.multiplatform.SentryEvent
 
 internal fun CocoaSentryEvent.applyKmpEvent(kmpEvent: SentryEvent): CocoaSentryEvent {
+    // Native SDKs may have a default behaviour if no release is set.
+    // Setting a release (even if it's null) will remove this behaviour,
+    // so we have to check if the values differ first before setting.
+    if (releaseName != kmpEvent.release) {
+        releaseName = kmpEvent.release
+    }
     kmpEvent.level?.let { level = it.toCocoaSentryLevel() }
     kmpEvent.platform?.let { platform = it }
     message = kmpEvent.message?.toCocoaMessage()
     logger = kmpEvent.logger
     fingerprint = kmpEvent.fingerprint
-    releaseName = kmpEvent.release
     environment = kmpEvent.environment
     user = kmpEvent.user?.toCocoaUser()
     serverName = kmpEvent.serverName