+++ sentry-kotlin-multiplatform/src/commonJvmMain/kotlin/io/sentry/kotlin/multiplatform/extensions/SentryEventExtensions.jvm.kt
@@ -4,12 +4,18 @@ import io.sentry.kotlin.multiplatform.JvmSentryEvent
 import io.sentry.kotlin.multiplatform.JvmSentryId
 import io.sentry.kotlin.multiplatform.SentryEvent
 
+// TODO(buenaflor): properly test this once we introduce a mocking framework
 internal fun JvmSentryEvent.applyKmpEvent(kmpEvent: SentryEvent): JvmSentryEvent {
+    // Native SDKs may have a default behaviour if no release is set.
+    // Setting a release (even if it's null) will remove this behaviour,
+    // so we have to check if the values differ first before setting.
+    if (release != kmpEvent.release) {
+        release = kmpEvent.release
+    }
     level = kmpEvent.level?.toJvmSentryLevel()
     message = kmpEvent.message?.toJvmMessage()
     logger = kmpEvent.logger
     fingerprints = kmpEvent.fingerprint
-    release = kmpEvent.release
     environment = kmpEvent.environment
     platform = kmpEvent.platform
     user = kmpEvent.user?.toJvmUser()